+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## [1.0.3]
+### Changed
+- The project no longer relies on gen-class, instead opting to have a JVM-only dependency.
+
+### Fixed
+- The `:no-error` clause in `handler-case` was called even on an error
+
 ## [1.0.2]
 ### Fixed
 - Exceptions were caught and converted to conditions around restarts
@@ -20,6 +27,8 @@ All notable changes to this project will be documented in this file. This change
 - Specs to all public functions
 - Basic implementation of conditions and restarts
 
+[Unreleased]: https://github.com/IGJoshua/farolero/compare/v1.0.3...HEAD
+[1.0.3]: https://github.com/IGJoshua/farolero/compare/v1.0.2...v1.0.3
 [1.0.2]: https://github.com/IGJoshua/farolero/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/IGJoshua/farolero/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/IGJoshua/farolero/compare/e2f23793cbf91f7c6dc35e61028bd99c4578bb4a...v1.0.0